backend/users/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
requests
pydantic[email]
pydantic-settings
